Implement routing, add new pages, and refactor directory
This commit sets up routing using react-router-dom and adds SignIn, Register, Home, and Detail pages. The App component is updated to route these pages. Also, the spelling mistake in the directory was corrected ("componenst" to "component") and the file structure was reorganized for better organization and readability.
