+++ package.json
@@ -14,6 +14,7 @@
     "antd": "^5.12.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.1",
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.6",
     "typescript": "^4.9.5",
