+++ src/component/Carousel/index.ts
