+++ src/component/Footer/index.ts
