+++ src/component/pages/detail/index.ts
@@ -1,0 +1,1 @@
+export * from "./Detail";
