+++ src/component/pages/home/mockups.ts
