+++ src/component/pages/index.ts
@@ -1,0 +1,4 @@
+export * from "./home";
+export * from "./signIn";
+export * from "./register";
+export * from "./detail";
