+++ src/component/pages/register/Register.tsx
@@ -1,0 +1,3 @@
+export const Register = () => {
+  return <>注册页面</>;
+};
