+++ src/component/pages/register/index.ts
@@ -1,0 +1,1 @@
+export * from "./Register";
