+++ src/component/pages/signIn/SignInPage.tsx
@@ -1,0 +1,3 @@
+export const SignInPage = () => {
+  return <>登录页面</>;
+};
